refactor: add-to-context typegen on build
This refactor removes the last place where addToContext type extraction
was integrated outside the new lib module. There was a wacky environment
variable system in use for passing the data at build time from build
module to nexus schema config module. Now the system used at build is
the same one used in dev!
